backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-nested-routers==0.93.4
django-filter==23.5
drf-yasg==1.21.7
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
django-redis==5.4.0
Pillow==10.1.0
python-decouple==3.8
drf-spectacular==0.26.5
django-storages==1.14.2
boto3==1.29.7
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
aiohttp==3.9.1
cryptography==41.0.7
pyotp==2.9.0
qrcode==7.4.2
django-otp==1.3.0
django-two-factor-auth==1.15.5
django-allauth==0.57.0
requests==2.31.0
httpx==0.25.1
python-jose==3.3.0
stripe==7.4.0
reportlab==4.0.7
weasyprint==60.1
jdatetime==4.1.1
python-dateutil==2.8.2
django-admin-rangefilter==0.11.2
django-import-export==3.3.3
django-extensions==3.2.3
django-debug-toolbar==4.2.0
gunicorn==21.2.0
uvicorn==0.22.0
whitenoise==6.6.0
sentry-sdk==1.38.0
prometheus-client==0.19.0
django-prometheus==2.3.1
user-agents==2.2.0
django-jazzmin==2.6.0
kavenegar==1.1.2

# Payment integration
